Splitting Drag and drop handler into separate component
