+++ src/components/Workspace.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { inject, observer } from 'mobx-react';
-import { DragDropContext } from 'react-beautiful-dnd';
+import DragDropMaster from './DragDropMaster';
 import { Plan } from './Plan';
 import { CourseModal } from './editCourseModal/CourseModal';
 import { Trash } from './Trash';
@@ -10,13 +10,11 @@ import '../styles/utilities.grab.scss';
 export const Workspace = inject('store')(observer(({ store }) => {
   return (
     <div className="workspace">
-      <DragDropContext
-        onDragEnd={store.mainPlan.handleDragDrop.bind(store.mainPlan)}
-      >
+      <DragDropMaster plan={store.mainPlan}>
         <Plan plan={store.mainPlan} />
         <CourseModal />
         <Trash />
-      </DragDropContext>
+      </DragDropMaster>
     </div>
   );
 }));
